refactor(desktop): split load session hydration stages
