+++ apps/desktop/src/state/operations/agent-orchestrator/lifecycle/load-sessions-stages.test.ts
@@ -1,0 +1,209 @@
+import { describe, expect, test } from "bun:test";
+import type { AgentSessionRecord } from "@openducktor/contracts";
+import type { AgentSessionState } from "@/types/agent-orchestrator";
+import {
+  hydrateSessionRecordsStage,
+  preparePersistedSessionMergeStage,
+  type SessionLoadIntent,
+} from "./load-sessions-stages";
+
+const createSession = (overrides: Partial<AgentSessionState> = {}): AgentSessionState => ({
+  sessionId: "session-1",
+  externalSessionId: "external-1",
+  taskId: "task-1",
+  role: "build",
+  scenario: "build_implementation_start",
+  status: "idle",
+  startedAt: "2026-03-01T09:00:00.000Z",
+  runtimeKind: "opencode",
+  runtimeId: null,
+  runId: null,
+  runtimeEndpoint: "",
+  workingDirectory: "/tmp/repo/worktree",
+  messages: [],
+  draftAssistantText: "",
+  draftAssistantMessageId: null,
+  draftReasoningText: "",
+  draftReasoningMessageId: null,
+  contextUsage: null,
+  pendingPermissions: [],
+  pendingQuestions: [],
+  todos: [],
+  modelCatalog: null,
+  selectedModel: null,
+  isLoadingModelCatalog: false,
+  promptOverrides: {},
+  ...overrides,
+});
+
+const createRecord = (overrides: Partial<AgentSessionRecord> = {}): AgentSessionRecord => ({
+  sessionId: "session-1",
+  externalSessionId: "external-1",
+  role: "build",
+  scenario: "build_implementation_start",
+  startedAt: "2026-03-01T09:00:00.000Z",
+  workingDirectory: "/tmp/repo/worktree",
+  runtimeKind: "opencode",
+  selectedModel: null,
+  ...overrides,
+});
+
+const createIntent = (overrides: Partial<SessionLoadIntent> = {}): SessionLoadIntent => ({
+  repoPath: "/tmp/repo",
+  taskId: "task-1",
+  mode: "bootstrap",
+  requestedSessionId: null,
+  requestedHistoryKey: null,
+  shouldHydrateRequestedSession: false,
+  shouldReconcileLiveSessions: false,
+  historyPolicy: "none",
+  ...overrides,
+});
+
+const createStateHarness = (sessions: Record<string, AgentSessionState>) => {
+  let state = sessions;
+  const sessionsRef = { current: state };
+  return {
+    sessionsRef,
+    setSessionsById: (
+      updater:
+        | Record<string, AgentSessionState>
+        | ((current: Record<string, AgentSessionState>) => Record<string, AgentSessionState>),
+    ) => {
+      state = typeof updater === "function" ? updater(state) : updater;
+      sessionsRef.current = state;
+    },
+    updateSession: (
+      sessionId: string,
+      updater: (current: AgentSessionState) => AgentSessionState,
+    ) => {
+      const current = state[sessionId];
+      if (!current) {
+        return;
+      }
+      state = {
+        ...state,
+        [sessionId]: updater(current),
+      };
+      sessionsRef.current = state;
+    },
+    getState: () => state,
+  };
+};
+
+describe("load-sessions-stages", () => {
+  test("uses the in-memory requested session record without reloading persisted sessions", async () => {
+    const existingSession = createSession({ runtimeEndpoint: "http://127.0.0.1:4444" });
+    const stateHarness = createStateHarness({ "session-1": existingSession });
+    let persistedLoads = 0;
+    let setCalls = 0;
+
+    const output = await preparePersistedSessionMergeStage({
+      intent: createIntent({
+        mode: "requested_history",
+        requestedSessionId: "session-1",
+        requestedHistoryKey: "/tmp/repo::task-1::session-1",
+        shouldHydrateRequestedSession: true,
+        historyPolicy: "requested_only",
+      }),
+      sessionsRef: stateHarness.sessionsRef,
+      setSessionsById: (updater) => {
+        setCalls += 1;
+        stateHarness.setSessionsById(updater);
+      },
+      isStaleRepoOperation: () => false,
+      loadPersistedRecords: async () => {
+        persistedLoads += 1;
+        return [createRecord()];
+      },
+      loadRepoPromptOverrides: async () => ({}),
+    });
+
+    expect(persistedLoads).toBe(0);
+    expect(setCalls).toBe(0);
+    expect(output.recordsToHydrate).toHaveLength(1);
+    expect(output.recordsToHydrate[0]?.sessionId).toBe("session-1");
+    expect(output.historyHydrationSessionIds.has("session-1")).toBe(true);
+  });
+
+  test("merges persisted records while preserving in-memory pending input", async () => {
+    const existingSession = createSession({
+      pendingPermissions: [{ requestId: "perm-current", permission: "read", patterns: [".env"] }],
+      pendingQuestions: [
+        {
+          requestId: "question-current",
+          questions: [{ header: "Confirm", question: "Ship it?", options: [] }],
+        },
+      ],
+    });
+    const stateHarness = createStateHarness({ "session-1": existingSession });
+
+    await preparePersistedSessionMergeStage({
+      intent: createIntent(),
+      sessionsRef: stateHarness.sessionsRef,
+      setSessionsById: stateHarness.setSessionsById,
+      isStaleRepoOperation: () => false,
+      loadPersistedRecords: async () => [
+        createRecord({
+          startedAt: "2026-03-01T10:00:00.000Z",
+          workingDirectory: "/tmp/repo/updated-worktree",
+        }),
+      ],
+      loadRepoPromptOverrides: async () => ({}),
+    });
+
+    const nextSession = stateHarness.getState()["session-1"];
+    expect(nextSession?.startedAt).toBe("2026-03-01T10:00:00.000Z");
+    expect(nextSession?.pendingPermissions).toEqual(existingSession.pendingPermissions);
+    expect(nextSession?.pendingQuestions).toEqual(existingSession.pendingQuestions);
+  });
+
+  test("marks requested-history hydration failed when runtime resolution fails", async () => {
+    const stateHarness = createStateHarness({ "session-1": createSession() });
+    let promptLoads = 0;
+
+    await expect(
+      hydrateSessionRecordsStage({
+        adapter: {
+          hasSession: () => false,
+          listLiveAgentSessionSnapshots: async () => [],
+          loadSessionHistory: async () => [],
+          resumeSession: async (input) => ({
+            sessionId: input.sessionId,
+            externalSessionId: input.externalSessionId,
+            role: input.role,
+            scenario: input.scenario,
+            startedAt: "2026-03-01T09:00:00.000Z",
+            status: "idle",
+            runtimeKind: input.runtimeKind,
+          }),
+        },
+        setSessionsById: stateHarness.setSessionsById,
+        updateSession: stateHarness.updateSession,
+        isStaleRepoOperation: () => false,
+        recordsToHydrate: [createRecord()],
+        historyHydrationSessionIds: new Set(["session-1"]),
+        runtimePlanner: {
+          readCurrentHydratedRuntimeResolution: () => null,
+          resolveHydrationRuntime: async () => ({
+            ok: false,
+            runtimeKind: "opencode",
+            reason: "No live runtime found for working directory /tmp/repo/worktree.",
+          }),
+          loadLiveAgentSessionSnapshot: async () => null,
+        },
+        promptAssembler: {
+          buildHydrationPreludeMessages: async () => [],
+          buildHydrationSystemPrompt: async () => "",
+        },
+        getRepoPromptOverrides: async () => {
+          promptLoads += 1;
+          return {};
+        },
+      }),
+    ).rejects.toThrow("No live runtime found for working directory /tmp/repo/worktree.");
+
+    expect(promptLoads).toBe(0);
+    expect(stateHarness.getState()["session-1"]?.historyHydrationState).toBe("failed");
+  });
+});
